requirements.txt
django
pyjwt
pillow
mysqlclient
djangorestframework
django-cors-headers
drf-yasg
django-prometheus
django-baton
